refactor(client): adjust file structure to match urls
