+++ client/src/views/components/config/editorFieldSettings.tsx
