+++ client/src/views/components/config/typeSettings/index.ts
