+++ client/src/views/components/config/typeSettings/select.tsx
