+++ client/src/views/components/config/typeSettings/subContent.tsx
