+++ client/src/views/components/config/typeSettings/typeOptionProps.ts
