+++ client/src/views/components/elements/editor/typeEditors/text.tsx
